add calc_measure function
This function generalises Megan's two prototype functions (calc_multi_measure and calc_onee_question_measure), converting it to just a single function.

Note that the function can take on n number of components, and that n = 1 should not be a special case that warrants a separate function.

Function also modified to return a list, rather than a named series. This removed the need for a separate argument for output_name in the function, only then to assign it to a named column.